feat: Added more info to the Movie class and changed style of the webpage to show the new info, also added README file
The Movie class was updated to get info about the year and director of the movie. The webpage html and css were updated using a Bootstrap template to show more info about the movie, like the year, the director and the storyline. For better viewing, a JQuery function was used to toggle the director and storyline of the movie when hovered. And a README file with instructions to run the code was added.